Added LocationString opaque type to enforce parseing of locations in one file.
