+++ meteor-app/imports/api/models/location-queries.js
@@ -1,0 +1,41 @@
+// @flow
+import { execTransactionRO } from "/imports/api/connectors/postgresql.js";
+
+import type { Location, Company, JobAd } from ".";
+import { parseLocationString } from ".";
+
+export async function getLocationsByCompany(
+	company: Company
+): Promise<Location[]> {
+	const transaction = async client => {
+		let locationResults = { rows: [] };
+
+		locationResults = await client.query(
+			"SELECT * FROM company_locations WHERE companyid=$1",
+			[company.companyid]
+		);
+
+		return locationResults.rows.map(loc =>
+			parseLocationString(loc.companylocation)
+		);
+	};
+
+	return execTransactionRO(transaction);
+}
+
+export async function getLocationsByJobAd(jobAd: JobAd): Promise<Location[]> {
+	const transaction = async client => {
+		let locationResults = { rows: [] };
+
+		locationResults = await client.query(
+			"SELECT * FROM job_locations WHERE jobadid=$1",
+			[jobAd.jobadid]
+		);
+
+		return locationResults.rows.map(loc =>
+			parseLocationString(loc.joblocation)
+		);
+	};
+
+	return execTransactionRO(transaction);
+}
